--- app/src/main/java/com/github/sdpsharelook/storage/RealTimeDatabaseRepository.kt
@@ -1,10 +1,0 @@
-package com.github.sdpsharelook.storage
-
-import kotlinx.coroutines.flow.Flow
-
-/**
- * Real time database using repository design pattern.
- */
-interface RealTimeDatabaseRepository {
-    fun fetchValues(): Flow<Result<Any?>>
-}
